+++ app/src/main/java/project/moms/assistant/data/api/Api.kt
@@ -1,0 +1,4 @@
+package project.moms.assistant.data.api
+
+class Api {
+}
