+++ app/src/main/java/project/moms/assistant/data/repository/Repo.kt
@@ -1,0 +1,4 @@
+package project.moms.assistant.data.repository
+
+class Repo {
+}
