+++ app/src/main/java/project/moms/assistant/presentation/adapter/Adapter.kt
@@ -1,0 +1,4 @@
+package project.moms.assistant.presentation.adapter
+
+class Adapter {
+}
