+++ app/src/main/java/project/moms/assistant/presentation/fragments/FragmentDiary.kt
@@ -1,0 +1,19 @@
+package project.moms.assistant.presentation.fragments
+
+import android.os.Bundle
+import androidx.fragment.app.Fragment
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import project.moms.assistant.R
+
+
+class FragmentDiary : Fragment() {
+
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        return inflater.inflate(R.layout.fragment_diary, container, false)
+    }
+}
