+++ app/src/main/res/layout/fragment_children_account.xml
@@ -3,6 +3,6 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".presentation.FragmentChildrenAccount">
+    tools:context=".presentation.fragments.FragmentChildrenAccount">
 
 </FrameLayout>
