+++ app/src/main/res/layout/fragment_main_screen.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".presentation.FragmentMainScreen">
+    tools:context=".presentation.fragments.FragmentMainScreen">
     <ScrollView
         android:id="@+id/scrollViewContent"
         android:layout_width="match_parent"
