+++ app/src/main/res/layout/fragment_sleep.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".presentation.FragmentSleep">
+    tools:context=".presentation.fragments.FragmentSleep">
     <ScrollView
         android:id="@+id/scrollViewContent"
         app:layout_constraintTop_toTopOf="parent"
