+++ app/src/main/res/navigation/nav_graph.xml
@@ -7,7 +7,7 @@
 
     <fragment
         android:id="@+id/fragmentAssistant"
-        android:name="project.moms.assistant.presentation.FragmentAssistant"
+        android:name="project.moms.assistant.presentation.fragments.FragmentAssistant"
         android:label="fragment_assistant"
         tools:layout="@layout/fragment_assistant" >
         <action
@@ -20,14 +20,16 @@
             android:id="@+id/action_fragmentAssistant_to_fragmentDiary"
             app:destination="@id/fragmentDiary" />
     </fragment>
+
     <fragment
         android:id="@+id/fragmentChildrenAccount"
-        android:name="project.moms.assistant.presentation.FragmentChildrenAccount"
+        android:name="project.moms.assistant.presentation.fragments.FragmentChildrenAccount"
         android:label="fragment_children_account"
         tools:layout="@layout/fragment_children_account" />
+
     <fragment
         android:id="@+id/fragmentDiary"
-        android:name="project.moms.assistant.presentation.FragmentDiary"
+        android:name="project.moms.assistant.presentation.fragments.FragmentDiary"
         android:label="fragment_diary"
         tools:layout="@layout/fragment_diary" >
         <action
@@ -40,9 +42,10 @@
             android:id="@+id/action_fragmentDiary_to_fragmentAssistant"
             app:destination="@id/fragmentAssistant" />
     </fragment>
+
     <fragment
         android:id="@+id/fragmentMainScreen"
-        android:name="project.moms.assistant.presentation.FragmentMainScreen"
+        android:name="project.moms.assistant.presentation.fragments.FragmentMainScreen"
         android:label="fragment_main_screen"
         tools:layout="@layout/fragment_main_screen" >
         <action
@@ -58,9 +61,10 @@
             android:id="@+id/action_fragmentMainScreen_to_fragmentChildrenAccount"
             app:destination="@id/fragmentChildrenAccount" />
     </fragment>
+
     <fragment
         android:id="@+id/fragmentSleep"
-        android:name="project.moms.assistant.presentation.FragmentSleep"
+        android:name="project.moms.assistant.presentation.fragments.FragmentSleep"
         android:label="fragment_sleep"
         tools:layout="@layout/fragment_sleep" >
         <action
@@ -73,4 +77,5 @@
             android:id="@+id/action_fragmentSleep_to_fragmentAssistant"
             app:destination="@id/fragmentAssistant" />
     </fragment>
+
 </navigation>
